+++ src/main/java/net/whydah/identity/user/search/Paginator.java
@@ -5,7 +5,7 @@ public class Paginator {
 
     static int pageSize = LuceneUserSearch.MAX_HITS;
 
-    public static ArrayLocation calculateArrayLocation(int totalHits, int pageNumber) {
+    public static ArrayLocation calculateArrayLocation(long totalHits, int pageNumber) {
         ArrayLocation al = new ArrayLocation();
  
         if (totalHits < 1 || pageNumber < 1 || pageSize < 1) {
@@ -15,7 +15,7 @@ public class Paginator {
         }
  
         int start= 1 + (pageNumber -1) * pageSize;
-        int end = Math.min(pageNumber * pageSize, totalHits);
+        int end = Math.min(pageNumber * pageSize, Math.toIntExact(totalHits));
         if (start > end) {
             start = Math.max(1, end - pageSize);
         }
